split utils.py into multiple modules; working on factor analysis
