+++ applications/forny_2023/eda.py
@@ -1,8 +1,5 @@
-from copy import deepcopy
-from joblib import Parallel, delayed
 import os
-from typing import List, Optional, Tuple, Dict, Union, Any
-import warnings
+from typing import List, Optional, Tuple, Dict, Union
 
 from anndata import AnnData
 from matplotlib.colors import Normalize
@@ -12,493 +9,13 @@ from mudata import MuData
 import numpy as np
 import pandas as pd
 import scanpy as sc
-import scipy.sparse as sp
-from scipy.stats import spearmanr, pearsonr, kstest
+from scipy.stats import spearmanr, pearsonr
 import seaborn as sns
-import statsmodels.api as sm
 
 
 # Set up the plotting style with a standard matplotlib style
 plt.style.use("default")
 sc.settings.set_figure_params(dpi=100, frameon=True)
-
-
-def plot_feature_counts_histogram(
-    adata: AnnData,
-    min_counts: int = 10,
-    layer: Optional[str] = None,
-    log_scale: bool = True,
-    n_bins: int = 50,
-    figsize: Tuple[int, int] = (10, 6),
-    save: Optional[str] = None,
-    show: bool = True,
-    title: Optional[str] = None,
-    xlim: Optional[Tuple[float, float]] = None,
-) -> Tuple[plt.Figure, plt.Axes]:
-    """
-    Plot a histogram of total counts per feature across all samples with
-    equally spaced bins on the transformed axis.
-    
-    Parameters
-    ----------
-    adata: anndata.AnnData
-        The AnnData object to analyze
-    min_counts: int, default=10
-        Threshold to mark with a vertical line
-    layer: str, optional
-        If provided, use this layer instead of .X
-    log_scale: bool, default=True
-        Whether to use log scale for x-axis
-    n_bins: int, default=50
-        Number of bins for the histogram
-    figsize: tuple, default=(10, 6)
-        Figure size as (width, height) in inches
-    save: str, optional
-        Path to save the figure
-    show: bool, default=True
-        Whether to display the figure
-    title: str, optional
-        Custom title for the plot
-    xlim: tuple, optional
-        Custom x-axis limits as (min, max)
-    
-    Returns
-    -------
-    Tuple containing:
-        - matplotlib.figure.Figure: The figure object
-        - matplotlib.axes.Axes: The axes object
-    """
-    # Select the matrix to use (X or a specific layer)
-    if layer is None:
-        matrix = adata.X
-        matrix_name = "X"
-    else:
-        if layer not in adata.layers:
-            raise ValueError(f"Layer '{layer}' not found in AnnData object")
-        matrix = adata.layers[layer]
-        matrix_name = layer
-    
-    # Calculate the total counts per feature (gene/transcript)
-    if sp.issparse(matrix):
-        feature_counts = np.array(matrix.sum(axis=0)).flatten()
-    else:
-        feature_counts = np.sum(matrix, axis=0)
-    
-    # Create the plot
-    fig, ax = plt.subplots(figsize=figsize)
-    
-    # Handle log scale with proper binning
-    if log_scale:
-        # Add a small pseudocount to avoid log(0)
-        pseudocount = 1
-        feature_counts_plus = feature_counts + pseudocount
-        min_counts_plus = min_counts + pseudocount
-        
-        # Determine minimum non-zero value for log scale
-        min_nonzero = np.min(feature_counts_plus[feature_counts_plus > 0])
-        
-        # Create log-spaced bins
-        if xlim is not None:
-            bin_min, bin_max = xlim
-        else:
-            bin_min = max(min_nonzero, 1)  # Ensure minimum is at least 1
-            bin_max = np.max(feature_counts_plus) * 1.01  # Add 1% to include maximum
-        
-        bins = np.logspace(np.log10(bin_min), np.log10(bin_max), n_bins)
-        
-        # Plot histogram with log-spaced bins
-        counts, bin_edges, _ = ax.hist(
-            feature_counts_plus, 
-            bins=bins, 
-            alpha=0.7, 
-            color='steelblue', 
-            edgecolor='black'
-        )
-        
-        # Set log scale for x-axis
-        ax.set_xscale('log')
-        x_label = "Total counts per feature (log scale)"
-        
-        # Add vertical line for min_counts threshold
-        ax.axvline(x=min_counts_plus, color='red', linestyle='--', linewidth=2, 
-                   label=f'Min counts threshold: {min_counts}')
-    else:
-        # For linear scale, use regular bins
-        if xlim is not None:
-            bin_min, bin_max = xlim
-        else:
-            bin_min = 0
-            bin_max = np.max(feature_counts) * 1.01  # Add 1% to include maximum
-        
-        bins = np.linspace(bin_min, bin_max, n_bins)
-        
-        # Plot histogram with linear-spaced bins
-        counts, bin_edges, _ = ax.hist(
-            feature_counts, 
-            bins=bins, 
-            alpha=0.7, 
-            color='steelblue', 
-            edgecolor='black'
-        )
-        
-        x_label = "Total counts per feature"
-        
-        # Add vertical line for min_counts threshold
-        ax.axvline(x=min_counts, color='red', linestyle='--', linewidth=2, 
-                   label=f'Min counts threshold: {min_counts}')
-    
-    # Count features below threshold
-    n_below_threshold = np.sum(feature_counts < min_counts)
-    percentage_below = (n_below_threshold / len(feature_counts)) * 100
-    
-    # Add annotations
-    ax.legend()
-    ax.set_xlabel(x_label)
-    ax.set_ylabel("Number of features")
-    
-    if title is None:
-        title = f"Distribution of total counts per feature ({matrix_name})"
-    title += f"\n{n_below_threshold} features ({percentage_below:.1f}%) have < {min_counts} counts"
-    ax.set_title(title)
-    
-    # Set x-axis limits if provided
-    if xlim is not None:
-        ax.set_xlim(xlim)
-    
-    # Add grid for better readability
-    ax.grid(True, alpha=0.3, which='both')
-    
-    plt.tight_layout()
-    
-    # Save figure if path is provided
-    if save is not None:
-        fig.savefig(save, bbox_inches='tight', dpi=300)
-    
-    # Show figure if requested
-    if show:
-        plt.show()
-    
-    return fig, ax
-
-
-def filter_features_by_counts(
-    adata: AnnData,
-    min_counts: int = 10,
-    layer: Optional[str] = None,
-    inplace: bool = True,
-) -> Optional[AnnData]:
-    """
-    Filter features to retain only those with more than min_counts reads across all samples.
-    
-    Parameters
-    ----------
-    adata: anndata.AnnData
-        The AnnData object to filter
-    min_counts: int, default=10
-        Minimum number of counts required for a feature to be kept
-    layer: str, optional
-        If provided, use this layer instead of .X for filtering
-    inplace: bool, default=True
-        Whether to modify the AnnData object in place or return a copy
-    
-    Returns
-    -------
-    anndata.AnnData, optional
-        Filtered AnnData object (only if inplace=False)
-    """
-    # Work with a copy if not inplace
-    if not inplace:
-        adata = adata.copy()
-    
-    # Select the matrix to use (X or a specific layer)
-    if layer is None:
-        matrix = adata.X
-    else:
-        if layer not in adata.layers:
-            raise ValueError(f"Layer '{layer}' not found in AnnData object")
-        matrix = adata.layers[layer]
-    
-    # Calculate the total counts per feature (gene/transcript)
-    if sp.issparse(matrix):
-        feature_counts = np.array(matrix.sum(axis=0)).flatten()
-    else:
-        feature_counts = np.sum(matrix, axis=0)
-    
-    # Create a mask for features with counts > min_counts
-    keep_features = feature_counts > min_counts
-    
-    # Apply the filter
-    if inplace:
-        adata._inplace_subset_var(keep_features)
-        filtered_adata = adata
-    else:
-        filtered_adata = adata[:, keep_features]
-    
-    # Store filtering information
-    n_features_before = adata.n_vars if not inplace else filtered_adata.n_vars + np.sum(~keep_features)
-    n_features_after = filtered_adata.n_vars
-    filtered_adata.uns["feature_filtering"] = {
-        "min_counts": min_counts,
-        "n_features_before": n_features_before,
-        "n_features_after": n_features_after,
-        "n_features_removed": n_features_before - n_features_after,
-    }
-    
-    if not inplace:
-        return filtered_adata
-
-
-def log2_transform(
-    adata: AnnData,
-    layer: Optional[str] = None,
-    inplace: bool = True,
-    pseudocount: float = 1.0,
-    create_layer: bool = True,
-    new_layer_name: Optional[str] = None,
-    modify_source: bool = False
-) -> Optional[AnnData]:
-    """
-    Log2-transforms the data matrix of an AnnData object.
-    
-    Parameters
-    ----------
-    adata: AnnData
-        The AnnData object to transform
-    layer: str, optional
-        If provided, transforms this layer instead of .X
-    inplace: bool, default=True
-        Whether to modify the AnnData object in place or return a copy
-    pseudocount: float, default=1.0
-        Value to add to the data before log-transforming to avoid log(0)
-    create_layer: bool, default=True
-        Whether to store the transformed data in a new layer
-    new_layer_name: str, optional
-        Name for the new layer containing transformed data.
-        If None, defaults to "log2" or "{layer}_log2"
-    modify_source: bool, default=False
-        Whether to modify the source data matrix. If False, the original
-        data will be preserved even when create_layer=True.
-    
-    Returns
-    -------
-    Optional[AnnData]
-        The log2-transformed object (only if inplace=False)
-    """
-    # Work with a copy if not inplace
-    if not inplace:
-        adata = adata.copy()
-    
-    # Function to transform a matrix
-    def transform_matrix(X: Union[np.ndarray, sp.spmatrix]) -> Union[np.ndarray, sp.spmatrix]:
-        """Transform a single data matrix using log2."""
-        if sp.issparse(X):
-            X_transformed = X.copy()
-            X_transformed.data = np.log2(X_transformed.data + pseudocount)
-            return X_transformed
-        else:
-            return np.log2(X + pseudocount)
-    
-    # Determine the source matrix to transform
-    if layer is None:
-        source_matrix = adata.X
-    else:
-        if layer not in adata.layers:
-            raise ValueError(f"Layer '{layer}' not found in AnnData object")
-        source_matrix = adata.layers[layer]
-    
-    # Determine the name for the new layer if creating one
-    if new_layer_name is None:
-        if layer is None:
-            new_layer_name = "log2"
-        else:
-            new_layer_name = f"{layer}_log2"
-    
-    # Create a new layer with transformed data if requested
-    if create_layer:
-        adata.layers[new_layer_name] = transform_matrix(source_matrix)
-    
-    # Only modify the source matrix if explicitly requested
-    if modify_source:
-        if layer is None:
-            if sp.issparse(adata.X):
-                adata.X.data = np.log2(adata.X.data + pseudocount)
-            else:
-                adata.X = np.log2(adata.X + pseudocount)
-        else:
-            if sp.issparse(adata.layers[layer]):
-                adata.layers[layer].data = np.log2(adata.layers[layer].data + pseudocount)
-            else:
-                adata.layers[layer] = np.log2(adata.layers[layer] + pseudocount)
-    
-    # Add a note in the var attributes to track the transformation
-    if create_layer:
-        adata.uns[f"{new_layer_name}_transform"] = {
-            "type": "log2",
-            "pseudocount": pseudocount,
-            "source": "X" if layer is None else layer
-        }
-    
-    if not inplace:
-        return adata
-
-
-def center_rows_and_columns(
-    adata: AnnData,
-    layer: Optional[str] = None,
-    max_iterations: int = 10,
-    tolerance: float = 1e-6,
-    inplace: bool = True,
-    copy_layer: bool = False,
-    new_layer_name: Optional[str] = None,
-) -> Optional[AnnData]:
-    """
-    Alternately center rows (features) and columns (cells) of an AnnData object
-    until both are approximately zero-centered.
-    
-    Parameters
-    ----------
-    adata: anndata.AnnData
-        The AnnData object to transform
-    layer: str, optional
-        If provided, transforms this layer instead of .X
-    max_iterations: int, default=10
-        Maximum number of iterations to perform
-    tolerance: float, default=1e-6
-        Convergence threshold for mean absolute deviation
-    inplace: bool, default=True
-        Whether to modify the AnnData object in place or return a copy
-    copy_layer: bool, default=False
-        Whether to create a new layer with the result
-    new_layer_name: str, optional
-        Name for the new layer containing transformed data.
-        If None but copy_layer is True, defaults to "centered" or "{layer}_centered"
-    
-    Returns
-    -------
-    anndata.AnnData, optional
-        The centered object (only if inplace=False)
-    """
-    # Work with a copy if not inplace
-    if not inplace:
-        adata = adata.copy()
-    
-    # Determine the source matrix and where to store the result
-    if layer is None:
-        source_matrix = adata.X
-    else:
-        if layer not in adata.layers:
-            raise ValueError(f"Layer '{layer}' not found in AnnData object")
-        source_matrix = adata.layers[layer]
-    
-    # Make a copy of the data matrix
-    if sp.issparse(source_matrix):
-        # For sparse matrices, we need to convert to dense as centering will likely
-        # destroy sparsity and lead to inefficient computations
-        matrix = source_matrix.toarray()
-    else:
-        matrix = source_matrix.copy()
-    
-    # Initialize variables for tracking convergence
-    n_rows, n_cols = matrix.shape
-    converged = False
-    
-    # Alternate centering
-    for i in range(max_iterations):
-        # Center rows (features)
-        row_means = np.mean(matrix, axis=1, keepdims=True)
-        matrix = matrix - row_means
-        
-        # Center columns (cells)
-        col_means = np.mean(matrix, axis=0, keepdims=True)
-        matrix = matrix - col_means
-        
-        # Check convergence
-        row_mean_abs = np.mean(np.abs(np.mean(matrix, axis=1)))
-        col_mean_abs = np.mean(np.abs(np.mean(matrix, axis=0)))
-        
-        if row_mean_abs < tolerance and col_mean_abs < tolerance:
-            converged = True
-            break
-    
-    # Warn if not converged
-    if not converged:
-        warnings.warn(
-            f"Centering did not converge after {max_iterations} iterations. "
-            f"Final mean absolute deviations: rows={row_mean_abs:.2e}, columns={col_mean_abs:.2e}. "
-            f"Consider increasing max_iterations or adjusting tolerance."
-        )
-    
-    # Store the result
-    if layer is None:
-        adata.X = matrix
-    else:
-        if copy_layer:
-            # Determine new layer name
-            if new_layer_name is None:
-                new_layer_name = f"{layer}_centered" if layer else "centered"
-            adata.layers[new_layer_name] = matrix
-        else:
-            adata.layers[layer] = matrix
-    
-    if not inplace:
-        return adata
-
-
-def center_rows_and_columns_mudata(
-    mdata,  # Not type-hinting to avoid import error for MuData
-    modalities: Optional[Union[str, List[str]]] = None,
-    layer: Optional[Union[str, Dict[str, str]]] = None,
-    max_iterations: int = 10,
-    tolerance: float = 1e-6,
-    inplace: bool = True,
-    copy_layer: bool = True,  # Changed default to True
-    new_layer_name: Optional[Union[str, Dict[str, str]]] = None,
-) -> Optional:
-    """
-    Apply alternate row-column centering to specified modalities in a MuData object.
-    """
-    # Work with a copy if not inplace
-    if not inplace:
-        mdata = deepcopy(mdata)
-    
-    # Determine which modalities to transform
-    if modalities is None:
-        modalities = list(mdata.mod.keys())
-    elif isinstance(modalities, str):
-        modalities = [modalities]
-    
-    # Handle layer parameter
-    if isinstance(layer, str) or layer is None:
-        layer_dict = {mod: layer for mod in modalities}
-    else:
-        layer_dict = layer
-    
-    # Handle new_layer_name parameter
-    if isinstance(new_layer_name, str) or new_layer_name is None:
-        new_layer_dict = {mod: new_layer_name for mod in modalities}
-    else:
-        new_layer_dict = new_layer_name
-    
-    # Apply centering to each modality
-    for mod in modalities:
-        if mod not in mdata.mod:
-            raise ValueError(f"Modality '{mod}' not found in MuData object")
-        
-        mod_layer = layer_dict.get(mod)
-        mod_new_layer = new_layer_dict.get(mod) if new_layer_dict else None
-        
-        center_rows_and_columns(
-            mdata.mod[mod],
-            layer=mod_layer,
-            max_iterations=max_iterations,
-            tolerance=tolerance,
-            inplace=True,  # Always inplace for the individual modality
-            copy_layer=copy_layer,  # This should be True to create new layers
-            new_layer_name=mod_new_layer
-        )
-    
-    if not inplace:
-        return mdata
 
 
 def plot_mudata_pca_variance(
@@ -1131,538 +648,6 @@ def plot_mudata_pcs(mdata: MuData,
                       fontsize=16, y=0.95, verticalalignment='top')
             figures[modality] = fig
     
-    return figures
-
-
-def apply_regression_per_feature(
-    adata: AnnData,
-    formula: str,
-    covariates_df: Optional[pd.DataFrame] = None,
-    layer: Optional[str] = None,
-    n_jobs: int = -1,
-    batch_size: int = 100,
-    progress_bar: bool = True
-) -> pd.DataFrame:
-    """
-    Apply linear regression to each feature in an AnnData object and return statistics.
-    Supports parallel processing across multiple cores using joblib.
-    
-    Parameters
-    ----------
-    adata : anndata.AnnData
-        The annotated data matrix.
-    formula : str
-        Formula for regression in patsy format (e.g., '~ batch + n_counts').
-        Don't include the dependent variable, as each feature will be used.
-    covariates_df : Optional[pd.DataFrame], optional
-        DataFrame containing covariates. If None, uses adata.obs.
-    layer : Optional[str], optional
-        If provided, use this layer instead of X. The layer can be a string referring
-        to a layer in adata.layers, or a key in adata.obsm for alternative feature matrices.
-    n_jobs : int, optional
-        Number of parallel jobs. -1 means using all processors.
-    batch_size : int, optional
-        Number of features to process in each batch.
-    progress_bar : bool, optional
-        Whether to display a progress bar.
-        
-    Returns
-    -------
-    pd.DataFrame
-        DataFrame with regression statistics for each feature.
-    """
-    # Set up covariates
-    if covariates_df is None:
-        covariates_df = adata.obs.copy()
-    
-    # Get feature names based on data source
-    if layer is None:
-        feature_names = adata.var_names
-    elif layer in adata.layers:
-        feature_names = adata.var_names
-    elif layer in adata.obsm:
-        # For obsm matrices, we need to generate feature names
-        if hasattr(adata.obsm[layer], 'columns'):
-            # If it's a DataFrame with column names
-            feature_names = adata.obsm[layer].columns
-        else:
-            # Otherwise, generate numeric feature names
-            n_features = adata.obsm[layer].shape[1]
-            feature_names = [f"{layer}_feature_{i}" for i in range(n_features)]
-    else:
-        raise ValueError(f"Layer '{layer}' not found in adata.layers or adata.obsm")
-    
-    # Get data to regress
-    if layer is None:
-        if isinstance(adata.X, np.ndarray):
-            X_data = adata.X
-        else:
-            X_data = adata.X.toarray()
-    elif layer in adata.layers:
-        if isinstance(adata.layers[layer], np.ndarray):
-            X_data = adata.layers[layer]
-        else:
-            X_data = adata.layers[layer].toarray()
-    elif layer in adata.obsm:
-        if isinstance(adata.obsm[layer], pd.DataFrame):
-            X_data = adata.obsm[layer].values
-        elif isinstance(adata.obsm[layer], np.ndarray):
-            X_data = adata.obsm[layer]
-        elif scipy.sparse.issparse(adata.obsm[layer]):
-            X_data = adata.obsm[layer].toarray()
-        else:
-            raise ValueError(f"Unsupported data type for obsm['{layer}']")
-    
-    # Define a function to process a single feature
-    def process_feature(i: int) -> List[Dict[str, Any]]:
-        feature_name = feature_names[i]
-        y = X_data[:, i]
-        
-        # Skip features with zero variance
-        if np.var(y) == 0:
-            return []
-        
-        # Create temporary dataframe with feature and covariates
-        temp_df = covariates_df.copy()
-        temp_df["y"] = y
-        
-        feature_results: List[Dict[str, Any]] = []
-        
-        try:
-            # Full formula with the feature as dependent variable
-            full_formula = "y " + formula
-            model = sm.formula.ols(formula=full_formula, data=temp_df)
-            result = model.fit()
-            
-            # Extract statistics for each coefficient
-            for coef_name in result.params.index:
-                if coef_name == "Intercept":
-                    continue  # Skip intercept
-                
-                feature_results.append(
-                    {
-                        "feature": feature_name,
-                        "coefficient": coef_name,
-                        "estimate": result.params[coef_name],
-                        "std_err": result.bse[coef_name],
-                        "statistic": result.tvalues[coef_name],
-                        "p_value": result.pvalues[coef_name],
-                        "conf_int_lower": result.conf_int().loc[coef_name, 0],
-                        "conf_int_upper": result.conf_int().loc[coef_name, 1],
-                    }
-                )
-        except Exception as e:
-            print(f"Error in feature {feature_name}: {str(e)}")
-        
-        return feature_results
-    
-    # Create feature indices
-    feature_indices: List[int] = list(range(X_data.shape[1]))
-    
-    # Verbose setting for progress bar
-    verbose = 10 if progress_bar else 0
-    
-    # Run parallel processing with joblib
-    print(f"Starting regression analysis with {n_jobs} cores...")
-    results_nested: List[List[Dict[str, Any]]] = Parallel(n_jobs=n_jobs, batch_size=batch_size, verbose=verbose)(
-        delayed(process_feature)(i) for i in feature_indices
-    )
-    
-    # Flatten results
-    results: List[Dict[str, Any]] = []
-    for sublist in results_nested:
-        results.extend(sublist)
-    
-    # Handle empty results
-    if not results:
-        print("No regression results were generated. Check your data and formula.")
-        return pd.DataFrame()
-    
-    # Convert to DataFrame
-    results_df = pd.DataFrame(results)
-    
-    # Add FDR correction (Benjamini-Hochberg)
-    if not results_df.empty and "p_value" in results_df.columns:
-        # Group by coefficient and calculate FDR-corrected p-values
-        coefs = results_df["coefficient"].unique()
-        for coef in coefs:
-            mask = results_df["coefficient"] == coef
-            results_df.loc[mask, "fdr_bh"] = sm.stats.multipletests(
-                results_df.loc[mask, "p_value"], method="fdr_bh"
-            )[1]
-    
-    print(f"Completed regression analysis for {len(results_df)} feature-coefficient pairs.")
-    return results_df
-
-
-def add_regression_results_to_anndata(
-    adata: AnnData,
-    results_df: pd.DataFrame,
-    key_added: str = "regression_results",
-    fdr_cutoff: float = 0.05,
-    inplace: bool = True,
-    effect_only_for_significant: bool = False,  # New parameter
-    add_all_stats: bool = True  # Add all statistics columns
-) -> Optional[AnnData]:
-    """
-    Add regression results dataframe to an AnnData object in a structured way.
-    
-    Parameters
-    ----------
-    adata : anndata.AnnData
-        The annotated data matrix.
-    results_df : pd.DataFrame
-        DataFrame with regression results as produced by apply_regression_per_feature().
-    key_added : str, optional
-        Key under which to add the results in adata.uns.
-    fdr_cutoff : float, optional
-        Significance threshold for FDR-corrected p-values. Default is 0.05.
-    inplace : bool, optional
-        If True, modify adata inplace. Otherwise return a copy. Default is True.
-    effect_only_for_significant : bool, optional
-        If True, effect sizes are only stored for significant features.
-        If False, effect sizes are stored for all features. Default is False.
-    add_all_stats : bool, optional
-        If True, add all available statistics (p-values, q-values, t-statistics, std errors)
-        to the adata.var dataframe. Default is True.
-        
-    Returns
-    -------
-    Optional[anndata.AnnData]
-        Depends on `inplace` parameter. If `inplace=True`, returns None,
-        otherwise returns a modified copy of the AnnData object.
-    """
-    if not inplace:
-        adata = adata.copy()
-    
-    # Make a copy to avoid modifying the original results
-    results = results_df.copy()
-    
-    # Create a hierarchical dictionary to store in adata.uns
-    regression_dict: Dict[str, Any] = {
-        "params": {
-            "features": results["feature"].unique().tolist(),
-            "coefficients": results["coefficient"].unique().tolist(),
-            "n_features": len(results["feature"].unique()),
-            "n_coefficients": len(results["coefficient"].unique()),
-            "fdr_cutoff": fdr_cutoff,
-            "timestamp": pd.Timestamp.now().strftime("%Y-%m-%d %H:%M:%S"),
-        },
-        "results": results.to_dict("records"),
-    }
-    
-    # Store in adata.uns
-    adata.uns[key_added] = regression_dict
-    
-    # Also annotate significant associations in var
-    # Create a significance mask for each coefficient
-    for coef in results["coefficient"].unique():
-        coef_safe = coef.replace(" ", "_").replace("-", "_")
-        
-        # Determine the p-value column to use (FDR-corrected if available)
-        p_col = "fdr_bh" if "fdr_bh" in results.columns else "p_value"
-        
-        # Get significant features for this coefficient
-        sig_mask = (results["coefficient"] == coef) & (results[p_col] < fdr_cutoff)
-        sig_features = results.loc[sig_mask, "feature"].unique()
-        
-        # Create a binary indicator in adata.var
-        adata.var[f"sig_{coef_safe}"] = adata.var_names.isin(sig_features)
-        
-        # Filter for the current coefficient
-        coef_results = results[results["coefficient"] == coef]
-        
-        # Create effect size dictionary for all features
-        effect_dict: Dict[str, float] = {}
-        for _, row in coef_results.iterrows():
-            effect_dict[row["feature"]] = row["estimate"]
-        
-        # Create effect size column for all features or only significant ones
-        if effect_only_for_significant:
-            # Original behavior: add effect sizes only for significant features
-            adata.var[f"effect_{coef_safe}"] = pd.Series(
-                [
-                    effect_dict.get(f, np.nan) if f in sig_features else np.nan
-                    for f in adata.var_names
-                ],
-                index=adata.var_names,
-            )
-        else:
-            # New behavior: add effect sizes for all features
-            adata.var[f"effect_{coef_safe}"] = pd.Series(
-                [effect_dict.get(f, np.nan) for f in adata.var_names],
-                index=adata.var_names,
-            )
-        
-        # Add all statistics as separate columns for all features
-        if add_all_stats:
-            # Raw p-values
-            pval_dict: Dict[str, float] = {}
-            for _, row in coef_results.iterrows():
-                pval_dict[row["feature"]] = row["p_value"]
-                
-            adata.var[f"pval_{coef_safe}"] = pd.Series(
-                [pval_dict.get(f, np.nan) for f in adata.var_names], 
-                index=adata.var_names
-            )
-            
-            # FDR-corrected p-values (q-values)
-            if "fdr_bh" in coef_results.columns:
-                qval_dict: Dict[str, float] = {}
-                for _, row in coef_results.iterrows():
-                    qval_dict[row["feature"]] = row["fdr_bh"]
-                    
-                adata.var[f"qval_{coef_safe}"] = pd.Series(
-                    [qval_dict.get(f, np.nan) for f in adata.var_names], 
-                    index=adata.var_names
-                )
-            
-            # T-statistics
-            tstat_dict: Dict[str, float] = {}
-            for _, row in coef_results.iterrows():
-                tstat_dict[row["feature"]] = row["statistic"]
-                
-            adata.var[f"tstat_{coef_safe}"] = pd.Series(
-                [tstat_dict.get(f, np.nan) for f in adata.var_names], 
-                index=adata.var_names
-            )
-            
-            # Standard errors
-            stderr_dict: Dict[str, float] = {}
-            for _, row in coef_results.iterrows():
-                stderr_dict[row["feature"]] = row["std_err"]
-                
-            adata.var[f"stderr_{coef_safe}"] = pd.Series(
-                [stderr_dict.get(f, np.nan) for f in adata.var_names], 
-                index=adata.var_names
-            )
-        else:
-            # Just add the selected p-value column as before
-            pval_dict: Dict[str, float] = {}
-            for _, row in coef_results.iterrows():
-                pval_dict[row["feature"]] = row[p_col]
-                
-            adata.var[f"pval_{coef_safe}"] = pd.Series(
-                [pval_dict.get(f, np.nan) for f in adata.var_names], 
-                index=adata.var_names
-            )
-    
-    # Add a summarized view for quick access to top features per coefficient
-    top_features: Dict[str, List[Dict[str, Any]]] = {}
-    for coef in results["coefficient"].unique():
-        # Sort by absolute effect size
-        coef_results = results[results["coefficient"] == coef].copy()
-        coef_results["abs_effect"] = np.abs(coef_results["estimate"])
-        coef_results = coef_results.sort_values("abs_effect", ascending=False)
-        
-        # Get top 50 features by effect size (with p-value < fdr_cutoff)
-        p_col = "fdr_bh" if "fdr_bh" in coef_results.columns else "p_value"
-        top = coef_results[coef_results[p_col] < fdr_cutoff].head(50)
-        
-        if len(top) > 0:
-            top_features[coef] = top[["feature", "estimate", p_col]].to_dict("records")
-    
-    # Store top features in uns
-    adata.uns[f"{key_added}_top_features"] = top_features
-    
-    return None if inplace else adata
-
-
-def plot_pvalue_histograms(
-    data: Union[AnnData, MuData],
-    regression_key: str = "regression_results",
-    output_dir: Optional[str] = None,
-    figsize: Tuple[int, int] = (12, 8),
-    include_stats: bool = True,
-    show_ks_test: bool = True,
-    fdr_cutoff: float = 0.05
-) -> Dict[str, Dict[str, plt.Figure]]:
-    """
-    Generate histograms of p-values and FDR-corrected p-values for all coefficients 
-    in regression results stored in AnnData or MuData objects.
-    
-    Parameters
-    ----------
-    data : Union[anndata.AnnData, mudata.MuData]
-        The annotated data matrix or multi-modal data object containing regression results.
-    regression_key : str, optional
-        Key under which regression results are stored in .uns. Default is "regression_results".
-    output_dir : Optional[str], optional
-        Directory to save histogram plots. If None, plots are displayed but not saved.
-    figsize : Tuple[int, int], optional
-        Figure size (width, height) in inches. Default is (12, 8).
-    include_stats : bool, optional
-        Whether to include summary statistics on the plots. Default is True.
-    show_ks_test : bool, optional
-        Whether to show Kolmogorov-Smirnov test results comparing p-value distribution
-        to the uniform distribution. Default is True.
-    fdr_cutoff : float, optional
-        Significance threshold for FDR-corrected p-values. Default is 0.05.
-        
-    Returns
-    -------
-    Dict[str, Dict[str, plt.Figure]]
-        Dictionary of generated figures indexed by modality and coefficient.
-    """
-    # Set plot style
-    sns.set_style("whitegrid")
-    plt.rcParams.update({'font.size': 12})
-    
-    # Initialize dictionary to store figure objects
-    figures: Dict[str, Dict[str, plt.Figure]] = {}
-    
-    # Function to process a single AnnData object
-    def process_anndata(
-        adata: AnnData, 
-        modality_name: Optional[str] = None
-    ) -> Dict[str, plt.Figure]:
-        # Check if regression results exist
-        if regression_key not in adata.uns:
-            print(f"No regression results found under key '{regression_key}'"
-                  f"{' in modality ' + modality_name if modality_name else ''}.")
-            return {}
-        
-        # Get regression results
-        regression_results = adata.uns[regression_key]
-        
-        # Convert results to DataFrame if needed
-        if isinstance(regression_results, dict) and "results" in regression_results:
-            results_df = pd.DataFrame(regression_results["results"])
-        elif isinstance(regression_results, pd.DataFrame):
-            results_df = regression_results
-        else:
-            print(f"Unexpected format for regression results" 
-                  f"{' in modality ' + modality_name if modality_name else ''}.")
-            return {}
-        
-        mod_figures: Dict[str, plt.Figure] = {}
-        
-        # Get unique coefficients
-        coefficients: List[str] = results_df["coefficient"].unique().tolist()
-        
-        for coef in coefficients:
-            # Filter results for current coefficient
-            coef_results = results_df[results_df["coefficient"] == coef]
-            
-            # Create figure - now with 2-3 panels depending on available data
-            n_panels = 3 if "fdr_bh" in coef_results.columns else 2
-            fig, axes = plt.subplots(1, n_panels, figsize=(figsize[0]*n_panels/2, figsize[1]))
-            fig.suptitle(f"P-value Distribution for Coefficient: {coef}"
-                         f"{' in ' + modality_name if modality_name else ''}", 
-                         fontsize=16)
-            
-            # Raw p-values
-            sns.histplot(coef_results["p_value"], bins=20, kde=True, ax=axes[0])
-            axes[0].set_title("Raw P-values")
-            axes[0].set_xlabel("P-value")
-            axes[0].set_ylabel("Count")
-            
-            # Add reference uniform distribution line
-            x = np.linspace(0, 1, 100)
-            y = len(coef_results) * 0.05  # Adjust height based on data
-            axes[0].plot(x, [y] * 100, 'r--', label='Uniform Distribution')
-            axes[0].legend()
-            
-            # Add summary statistics
-            if include_stats:
-                avg_p = coef_results["p_value"].mean()
-                median_p = coef_results["p_value"].median()
-                sig_count = sum(coef_results["p_value"] < 0.05)
-                sig_pct = 100 * sig_count / len(coef_results)
-                
-                stats_text = f"Mean: {avg_p:.4f}\nMedian: {median_p:.4f}\n"
-                stats_text += f"Significant: {sig_count}/{len(coef_results)} ({sig_pct:.1f}%)"
-                
-                axes[0].text(0.05, 0.95, stats_text, 
-                             transform=axes[0].transAxes, 
-                             verticalalignment='top',
-                             bbox=dict(boxstyle='round', facecolor='white', alpha=0.8))
-            
-            # QQ plot of p-values (to check for uniform distribution)
-            # This helps identify deviations from expected p-value distribution
-            pvals = np.sort(coef_results["p_value"])
-            expected = np.linspace(0, 1, len(pvals) + 1)[1:]
-            
-            axes[1].scatter(expected, pvals, alpha=0.5)
-            axes[1].plot([0, 1], [0, 1], 'r--')
-            axes[1].set_title("P-value QQ Plot")
-            axes[1].set_xlabel("Expected P-value")
-            axes[1].set_ylabel("Observed P-value")
-            
-            # FDR-corrected p-values (if available)
-            if "fdr_bh" in coef_results.columns:
-                sns.histplot(coef_results["fdr_bh"], bins=20, kde=True, ax=axes[2])
-                axes[2].set_title("FDR-corrected P-values (Benjamini-Hochberg)")
-                axes[2].set_xlabel("Adjusted P-value")
-                axes[2].set_ylabel("Count")
-                
-                # Add reference uniform distribution line
-                axes[2].plot(x, [y] * 100, 'r--', label='Uniform Distribution')
-                axes[2].legend()
-                
-                # Add summary statistics
-                if include_stats:
-                    avg_fdr = coef_results["fdr_bh"].mean()
-                    median_fdr = coef_results["fdr_bh"].median()
-                    sig_count_fdr = sum(coef_results["fdr_bh"] < fdr_cutoff)
-                    sig_pct_fdr = 100 * sig_count_fdr / len(coef_results)
-                    
-                    stats_text = f"Mean: {avg_fdr:.4f}\nMedian: {median_fdr:.4f}\n"
-                    stats_text += f"Significant: {sig_count_fdr}/{len(coef_results)} ({sig_pct_fdr:.1f}%)"
-                    
-                    axes[2].text(0.05, 0.95, stats_text, 
-                                 transform=axes[2].transAxes, 
-                                 verticalalignment='top',
-                                 bbox=dict(boxstyle='round', facecolor='white', alpha=0.8))
-            
-            # Calculate KS test against uniform distribution (expected for null hypothesis)
-            if show_ks_test:
-                ks_stat, ks_pval = kstest(coef_results["p_value"], 'uniform')
-                
-                # Add KS test result
-                fig.text(0.5, 0.01, 
-                         f"Kolmogorov-Smirnov test against uniform distribution: "
-                         f"statistic={ks_stat:.4f}, p-value={ks_pval:.4f}",
-                         ha='center', fontsize=12)
-            
-            plt.tight_layout(rect=[0, 0.03, 1, 0.95])
-            
-            # Save figure if output_dir is provided
-            if output_dir is not None:
-                # Create modality directory if needed
-                mod_dir = os.path.join(output_dir, 
-                                     modality_name if modality_name else "anndata")
-                os.makedirs(mod_dir, exist_ok=True)
-                
-                # Save figure
-                safe_coef = coef.replace(" ", "_").replace("/", "_").replace("\\", "_")
-                file_path = os.path.join(mod_dir, f"pvalue_hist_{safe_coef}.png")
-                fig.savefig(file_path, dpi=300, bbox_inches='tight')
-                print(f"Saved histogram to {file_path}")
-            
-            # Store figure
-            mod_figures[coef] = fig
-        
-        return mod_figures
-    
-    # Check if input is MuData or AnnData
-    if hasattr(data, 'mod'):  # MuData object
-        for modality in data.mod:
-            print(f"Processing {modality}...")
-            mod_figures = process_anndata(data[modality], modality_name=modality)
-            if mod_figures:
-                figures[modality] = mod_figures
-    else:  # AnnData object
-        figures['anndata'] = process_anndata(data)
-    
-    # Display figures if not saving
-    if output_dir is None:
-        for modality in figures:
-            for coef in figures[modality]:
-                plt.figure(figures[modality][coef].number)
-                plt.show()
-    
-    print("Histogram generation complete.")
     return figures
